+++ src/aioamazondevices/const.py
@@ -150,6 +150,14 @@ DEVICE_TO_IGNORE: list[str] = [
     "A2TF17PFR55MTB",  # Alexa App for Android
     "A1RTAM01W29CUP",  # Alexa App for PC
     "A18BI6KPKDOEI4",  # ecobee4 Smart Thermostat with Built-in Alexa - issue #199
+    "A21Z3CGI8UIP0F",  # Denon AVR-X1600H - issue #253
+    "A15ERDAKK5HQQG",  # unsupported Sonos devices - issue #257
+    "A3GZUE7F9MEB4U",  # Sony headset WH-1000XM3 - issue #269
+    "A23ZD3FSVQM5EE",  # Sony headset WH-1000XM2 - issue #326
+    "A7S41FQ5TWBC9",  # Sony headset WH-1000XM4 - issue #327
+    "A1L4KDRIILU6N9",  # Sony headset WH-CH700N  - issue #345
+    "A2IJJ9QXVOSYK0",  # JBL TUNE770NC - issue #391
+    "AKOAGQTKAS9YB",  # Amazon Echo Connect - issue #406
 ]
 
 DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
@@ -234,6 +242,11 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
         "model": "Echo Dot Clock",
         "hw_version": "Gen5",
     },
+    "A2F7IJUT32OLN4": {
+        "manufacturer": "Samsung Electronics Co., Ltd.",
+        "model": "Soundbar Q990D",
+        "hw_version": None,
+    },
     "A2GFL5ZMWNE0PX": {
         "model": "Fire TV",
         "hw_version": "Gen3",
@@ -262,6 +275,16 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
         "model": "Fire Tablet HD 10 Plus",
         "hw_version": "Gen11",
     },
+    "A2OSP3UA4VC85F": {
+        "manufacturer": "Sonos Inc.",
+        "model": "Sonos One",
+        "hw_version": "Gen1",
+    },
+    "A2RG3FY1YV97SS": {
+        "manufacturer": "Sonos Inc.",
+        "model": "Sonos Move",
+        "hw_version": "Gen1",
+    },
     "A2RU4B77X9R9NZ": {
         "model": "Echo Link Amp",
         "hw_version": None,
@@ -273,6 +296,11 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
     "A2UONLFQW0PADH": {
         "model": "Echo Show 8",
         "hw_version": "Gen3",
+    },
+    "A2Z8O30CD35N8F": {
+        "manufacturer": "Sonos Inc.",
+        "model": "Sonos Arc",
+        "hw_version": "Gen1",
     },
     "A303PJF6ISQ7IC": {
         "model": "Echo Auto",
@@ -381,6 +409,11 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
         "model": "Fire TV Stick",
         "hw_version": "Gen1",
     },
+    "AECNEXTDY5AD9": {
+        "manufacturer": "Cozyla",
+        "model": "Frame with Alexa",
+        "hw_version": None,
+    },
     "AIPK7MM90V7TB": {
         "model": "Echo Show 10",
         "hw_version": "Gen3",
@@ -397,6 +430,10 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
         "model": "Echo Stereo Pair",
         "hw_version": "Virtual",
     },
+    "AQ24620N8QD5Q": {
+        "model": "Echo Show 15",
+        "hw_version": "Gen2",
+    },
     "ASQZWP4GPYUT7": {
         "model": "Echo pop",
         "hw_version": "Gen1",
@@ -409,6 +446,11 @@ DEVICE_TYPE_TO_MODEL: dict[str, dict[str, str | None]] = {
         "manufacturer": "ecobee Inc.",
         "model": "ecobee Switch+",
         "hw_version": None,
+    },
+    "AVD3HM0HOJAAL": {
+        "manufacturer": "Sonos Inc.",
+        "model": "Sonos One",
+        "hw_version": "Gen2",
     },
     "AVU7CPPF2ZRAS": {
         "model": "Fire Tablet HD 8 Plus",